[ORC] Turn off dictionary encoding if the number of distinct strings is above threshold
Summary:
Add a configurable threshold so that if the number of distinct values in a string column is greater than that fraction of non-null values, dictionary encoding is turned off.

This is checked at the stripe level so within the same file the same column may be encoded differently in different stripes.

Modified the StringTreeReader to be able to handle different encodings in different stripes.

Test Plan:
Added a testcase to verify that dictionary encoding is turned off in certain cases using the file dump.

Added a testcase to verify that if dictionary encoding is turned off the data can still be read.

Added a testcase to verify that if dictionary encoding is turned on for some stripes and off for others for the same column in the same file, the data can still be read.