Integer, Short, Long dictionary encoding enabled
Summary:
Add support for Integer, Short, and Long dictionary encoding - based on
experimentation, dictionary encoding does as well if not better than RLE
for numeric types - as such, RLE will be replaced with dictionary
encoded Numeric columns. RLE will still be used to record string lengths
in String columns.

Adds a DynamicLongArray class which stores the keys in the Numeric
Dictionary - this is a memory efficient Long Array which is templated
off Dynamic Int Array.

Changes the RLEIntegerWriter to extend PositionedOutputStream since it is used
exclusively as a wrapper for writing RLE numeric values to a
PositionedOutputStream.

Dictionary Encoding Numeric Values is done by storing a memory efficient
and performant HashMap from the fastutils library. This library is
optimized for large collections and handles hash collisions with linear
probing. The Hashmap is used to track the unique set of keys in a
numeric column, and when the Numeric TreeWriter determines that the
cardinality:number of rows ratio is below a maximum threshold, the
unique keys are sorted (if specified as such) in order to improve
compression of the keys. This is a CPU performance optimization, which
guarantees that the nlogn performance hit is only encountered if
the column gets dictionary encoded and not otherwise.

Test Plan: added unit tests