Replace StringRedBlackTree with Hashmap implementation
Summary:
Removes StringRedBlackTree and replaces with hashmpa for Dictionary
Encoding. This change is a CPU optimization: the dictionary keys are no
longer sorted unless the query is configured to sort the dictionary
keys, minimizing the CPU cost to columns for which dictionary encoding
doesn't improve compression.

This new implementation of dictionary encoding stores the actual bytes
for a given row in a DynamicByteArray so that memory usage is explicit
and not left for the garbage collector to manage implicitly. The
fastutils hashmap used is a memory efficient implementation for large
collections, and tracks a logical representation of the text via its
real location in the DynamicByteArray as well as its hashcode.
Additionally, the hashmap uses a custom hashing strategy to determine if
there is a hash collision: two items are determined to be equal if the
bytes they store are actually equivalent.

In order to sort the keys in a memory efficient way, an array of
positional indices are sorted (this can be thought of as references into
the DynamicByteArray) using a custom comparator which compares the
actual bytes when sorting, but shuffles only the positional index
around.

The combination of these techniques yields a 1.4x memory efficiency
when compared to using Java hashmaps.

Test Plan: Unit tests