[ORC] Turn off dictionary encoding in presence of low entropy
Summary:
If the entropy is below a configured threshold (estimated as the number of distinct characters in a sample of strings from the dictionary) turn off dictionary encoding.

If the number of distinct strings is very low (by default <10% of all strings but it's configurable) keep dictionary encoding on.

If the entropy heuristic does not turn off dictionary encoding, use the existing metric of number of distinct strings to possibly turn it off.

For now, the heuristic is checked when writing a stripe out, and is rechecked for every stripe.  This can be optimized in the future.

Test Plan:
Added a testcase to verify that the heuristic can turn off dictionary encoding.

Since I didn't have to change existing tests, this shows that dictionary encoding is only turned off when appropriate.