[ORC] HIVE-4340 ORC should provide raw data size
Summary:
Pulling in the patch that adds raw data size to the ORC file format at the stripe level, and returns the raw data size at the row level to the FileSinkOperator for standard stats aggregation.

Added the raw data size at the file level to the file footer (this is not in the patch in open source)

Test Plan: Running ptests.