[ORC] Make vints optional
Summary:
This patch makes vints optional (by default they remain on).

SerializationUtils now takes the size of values in bytes and whether to use vints when reading and writing numerics.  Whether vints were used is included as an optional property of streams (again default is true).

Test Plan:
Added a testcase.

Tests had to be updated because of the extra handful of bytes it takes to store whether or not vints were used