[ORC] Make original ORC configs configurable via config file
Summary:
The original configs added to ORC (stripe size, compression window, compression method, etc.) were hardcoded to default values, and the only way to change them was to change the SerDe properties of individual tables.

With this change, it's possible to change the default values using Hive configs (in particular by changing hive-site.xml).

This will make it much easier to test different default values and change them if we find certain values are better in some clusters than others (e.g. coldstore vs. production).  This also allows us to store the default values as their appropriate types rather than strings.

The way the values are decided now is:
if the user set it in the SerDe properties, use that value
else if the value is set in the Hive configs, use that value
else use the default value

Test Plan: The tests pass.