[ORC] Remove unnecessary if statements from IntegerTreeWriter
Summary:
Looking at the write method for IntegerTreeWriter, I noticed that within a case statement 3 cases use the same code which is basically just an if block to break apart those 3 cases.  The if statements could be avoided by putting the appropriate code from each branch in the if statement into each of the 3 cases.

This improved the performance of the method (the most expensive method in ORC after anything in Dynamic arrays) by about 30%.

Test Plan: Ptests pass.