[ORC] Import Slice into Hive
Summary:
For ease of review, this diff only contains the files which I copied from Slice https://github.com/airlift/slice

The only changes I have made are to make it compatible with Java 6, add headers to files missing them, and to upgrade guava which is required to get these to compile.

While I am not using these in this diff, I have a couple diffs out which do.

Test Plan: Running ptests.