[DWRF] DynamicByteArray to use Slice instead of byte[][]
Summary:
Changed DynamicByteArray to use Slice instead of byte[][]

This improves the performance of compare by 20-30%.

Even though grow now requires copying memory, the performance of add in the profiles I looked at was actually slightly improved.

The only performance hit was in the constructor and clear, I assume because these now have to allocate an additional object, but the performance benefits far outway these.

I basically stole the Slice code from https://github.com/airlift/slice/tree/master/src/main/java/io/airlift/slice

I had to change it to make it compatible with Java 6, and made some changes to make it less generic and more specific to byte[]'s (and faster for our purposes :) )

Test Plan: Ran ptests and verified output on a couple partitions (we'll need to run snapshot tests on it with ORC on).