[ORC] Couple String dictionary optimizations
Summary:
In TextCompressedHashStrategy we were calling compare and checking that it equals 0.  This requires wasting a bunch of time figuring out which one is greater when they are not equal.  Changed it to use equals.

Changed it to check equality of lengths early on to save a bunch  of method calls.

Changed it to store the length in the TextCompressed object.  This saves a bunch of repeated calls to the getEnd method.

All in all this gets us a couple percentage points closer to RC.

Test Plan: Running ptests.