[ORC] compare hashes before comparing strings
Summary:
Apparently the fast utils method for adding strings to an open object custom hash set, is to take a hash of the hash to get the bucket the object belongs in.  It then compares the object to every other object in the bucket.  For some reason it doesn't bother to compare their original hash codes first, added that to the equals method for TextCompressedHashStrategy and got the benchmark ~5% faster.

Test Plan: Running ptests.