[DWRF] Don't allocate Text objects in StringDictionaryEncoder
Summary:
Right now, every time we check if a value exists in the dictionary/add a new value to the dictionary, we create a new Text object and containing the value were looking for/adding.  This involves allocating a new byte[] etc.  Since the object is never modified, it's safe to just use the Text object being passed in.

Also, when we sort the dictionary, every comparison involved in that sort allocates a new Text object to contain the bytes on one side of the comparison.  Since both sides of the comparison are in the same DynamicByteArray object, I added a new method which compares the bytes at 2 different offsets with 2 different lengths in the same DynamicByteArray to do this.  I also fixed Slice, so that comparing with a Slice, just calls compare on the underlying byte[] since the code is almost identical.

The perf gains from this are pretty much negligible, but it means we're allocating less memory, and I cleaned up some code, so I think it's a win.

Test Plan: Ran ptests, the snapshot 25 tables, and the benchmark.