[DWRF] Fix issues with RC table and ORC partition
Summary:
Based on my investigations, it looks like whoever wrote the code to convert ObjectInspectors from one format to another decided to only implement it for primitive types, moreover, they didn't even bother to make Unions convertable at all.  This is necessary if the table and partition are of different formats.

I fixed it so that for complex types, it converts the complex object inspector as well as the object inspectors of the types it contains.

I also made it so that unions can be converted, provided they are of either ORC's Serde, or standard union (what RC uses).  For the other types I wasn't sure what to do, so I'm going to leave them as is (other complex types have only some object inspectors which can be converted).

godeeper

Test Plan:
Created a test case with a table which is RC and a partition that is ORC the columns cover all the types I could think of.

Running ptests.