[DWRF] Use custom hash set in StringDictionaryEncoder
Summary:
Instead of using the fast utils hash set, I wrote one, which is based on it, but allows to avoid a lot of virtual calls, and object allocations.

Everything  is stored in 4 int arrays instead of an object per value.

The changes made on the original fast utils hash set were to support these 4 arrays instead of objects, I also added chaining instead of probing.  I noticed that we ended up with these blocks of values in the hash map, and we'd have to keep probing, but chaining reduced the number of comparisons needed.  I also modified the chaining, so if we try to add a value that already exists, it's moved to the head of the chain.  For dim_all_users, this seemed to help for columns where we were adding the same values over and over again, it didn't hurt the other case significantly.

Test Plan:
Ran ptests.

Ran it against the standard set of a few tables, the effect seemed to vary between none at all (but never worse), to 4-5% benefit.