[DWRF] Fix string dictionary size estimation
Summary: In the string dictionary, when we are estimating the size of the byte array used to store the strings, we use the amount of bytes of the array used, not the amount of bytes allocated for the array.  This leads to OOMs in snapshot testing.

Test Plan:
Ptests pass.

Verified this fixed a few of the ptests which OOM'd.