[DWRF] Make reading string dictionary more efficient
Summary:
Two fixes here that reduce the amount of memory needed to read in string dictionaries (fixes remaining OOMs in snapshot tests) and as a bonus also probably improves CPU perf (guessing, I haven't verified).

First, I changed the reader to read in the lengths before reading in the dictionary.  Since we sum up the lengths to get the offsets, we can use this to figure out the total amount of data in the dictionary stream.  Using this, we can initialize the dictionary DynamicByteArray to be the exact size we want, rather than starting with 256 Kb, and growing it as we read data.  This lead to OOMs because when we increase the size of the array we create a larger array and do a memcopy, during which time we are using potentially way more memory than the actual size of the dictionary.  (This is where I'm guessing the CPU savings come from because we're saving a bunch of GC's and memcopies).

Second, I changed the readAll in DynamicByteArray to end when in.available() == 0 instead of read == 0.  This is needed because if we allocate the perfect size, the last time data is read, read will be >0, and so it will loop through again, calling grow, only to then read 0 bytes.  This allows us to only use the exact byte size of the dictionary.

Test Plan:
Ran ptests, they pass.

Verified it fixed the remaining OOMs in snapshot tests.