[DWRF] move ORC merge code into new package
Summary:
The merge code for ORC is pretty Hive specific.  Concatenate is a Hive command that is never called from any other client.

By moving it to a separate package it's clear that this code is part of Hive and not part of the format itself.

I've also moved the testcases out of TestOrcFile (always a weird place for them) into their own test class.

This will help reduce the amount of shim code we'll have to provide clients in order to build against our jar.

Test Plan: Ran ptests.