[DWRF] change TestFileDump to use Snappy codec
Summary:
Running the tests via maven, and running the tests via ant, the compressed size of the data varies for this test alone.  I've verified that the data is exactly the same and the same number of rows are in each stripe.  It's just a few bytes smaller/larger in some streams.

I've tried switch between Java 6 and Java 7 in ant and the canonical results are produced either way.

I'm not sure what else to check, so I tried switching to Snappy and the results are the same across maven and ant.  (It was the same for no compression)  I suspect it has to do with the fact that we explicitly import specific versions of Snappy, rather than relying on a built in Java library like deflate.

The ZLIB code path still has more explicit tests in TestOrcFile to make sure it works, and the integration tests continue to use Zlib.

Test Plan: Ptests pass and the tests pass in maven.