[DWRF] Remove dependency on HiveConf from ReaderWriterProfiler
Summary:
The ORC code depends on ReaderWriterProfiler which is a new class we've added to internal Hive.  This means that in order for Presto to work with it we need to include it in the shims jar.  This means ReaderWriterProfiler can't depend on configs in HiveConf we've added that aren't in open source because Presto uses open source Hive.

I've added the hardcoded strings to the file, along with their defaults.  I've left them in HiveConf as well for when they're used in other classes (e.g. IOBenchmark) and also because it provides some nice type checking through the CLI.

Test Plan: ptests pass