[DWRF] Directly encoded string optimization
Summary: When the direct encoding for strings is carried over, we write the string to a byte array first, and then to a stream to be compressed in the write stripe stage.  The writing to the stream to be compressed can be done directly rather than going through the byte array first.  This also means that more data can be written into stripes since we store compressed data in memory, rather than the original uncompressed strings.

Test Plan: Ptests pass.