[DWRF] The NPE fix introduced an NPE
Summary:
When a table has more columns than a partition, the LazyOrcRow returns null for these additional columns.  getStructFieldData does not check for this before calling materialize

The number of branches in this new code should be the same as in the original code, just not in a single line.

Test Plan: The test which was broken is now fixed.