[DWRF] Index written incorrectly for numerics written directly
Summary:
If the direct encoding for numerics is chosen, then index is written out incorrectly for each stripe except the first stripe when first stripe carryover is used. Specifically the value 0 is stored as the beginning position of each index entry. This is caused by the use of the stream of indexes into the dictionary (which is not populated for direct encoding) to get the position. This breaks lazy reading of the column when values are skipped across index boundaries.

Fixed by using the stream data is written out to directly for the position of the index when direct encoding is used.

Plus some minor refactoring.

Test Plan:
Fixed TestOrcFile to demonstrate the issue.

Ptests pass.