[DWRF] Fix assumption that nextIsNull is called before materialize
Summary:
This diff contains a bunch of fixes mostly related to assumptions I had made that nextIsNull would be called before materialize which doesn't hold for group bys in Hive, and of course may not hold in other query engines.

Will describe them in comments on the diff.

Test Plan:
Added a bunch of new tests to demonstrate most if not all the issues.

Ptests pass.