[DWRF] Add field names to OrcStruct
Summary:
Added the field names to the OrcStruct object.  It's constant per file of input data (so doesn't consume tons of memory and doesn't require lots of allocations)  and is copied from the type metadata (the same way it gets passed to the object inspector).

This is a Presto request.

Test Plan:
ptests pass.

Added some tests for it to TestOrcFile.