[DWRF] Override hashCode in OrcLazyObject
Summary:
Overriding the hash code of OrcLazyObject similar to the way it is done for lazy objects from RC files.

This hash code is used by the GoupByOperator for hash aggregations.  The way it is now, it's just using the default method provided by Object, which, since we recycle objects, always puts the objects in the same hash bucket.  So the hash aggregation takes a very very long time.

Test Plan: Ran a group by query that was taking a long time before relative to RC, now it runs faster.