[DWRF] seek re-decompresses the same data repeatedly in some cases
Summary:
When reading lazily, if the reader needs to skip rows across index strides, it will seek to the start of the index stride and only skip rows in that stride.  If the same compression block of any stream used by this column is used after the seek, that compression block will be re-decompressed.

Fixed it so that seek called on an InStream will first check if it is in the same compression block, if so, it will only adjust it's position in the uncompressed data, rather than re-decompressing the data.

Test Plan:
Ptests pass.

On one of the snapshot queries that was showing particularly bad read performance this improved the speed recorded by ReaderWriterProfiler by 5x.

Also, as a bonus, TestOrcFile which used to run in ~2m20s now runs in about 1min15s.