[DWRF] Numeric dictionaries retaining memory after clear
Summary:
The dictionary used for numeric columns do not release the memory they hold after clear is called on them.  This means that after the dictionary is filled and the stripe is flushed, the next stripe can be flushed far too quickly because the estimated memory usage accounts for this large dictionary.

I came across this while attempting to run a microbenchmark on a table consisting of a single column of ~750,000,000 distinct bigints.  Since we default to dictionary encoding for the first stripe, the dictionary was filled, then each subsequent stripe contained only 5000 rows.

After this fix I saw a 1% decrease in the compressed size of the table, and a 25% reduction in the CPU (that includes both the time to read the data from the RC file and the time to write it to the ORC format).  I ran the benchmark 5 times to be sure.

Test Plan: Running ptests.