ReaderWriterProfiler to provide exclusive total read/write times
Summary:
The ReaderWriterProfiler currently assumes that are counters are measured exclusively, however this is not the case and it is a very easy mistake to make.

To fix this, I assigned a type to each counter (either read or write) and measure the total read/write time through counters for these types.  This way if two counters for different steps overlap (e.g. encoding and compressing) the read time will not double count the time.

Test Plan:
Ran this on a sample ORC and a sample RC query.  It showed that the write time for ORC double counts the majority of the compression and serialization time in the encoding time.

Ptests pass