Set deployAtEnd to true
Summary:

Our tests seem to have some order dependency which causes them to fail
on hive-dwrf about half the time.  Given the hive-dwrf-shims is built
and deployed before hive-dwrf this causes us to repdeploy it a lot.
Setting deployAtEnd to true, so that the packages are only depoyed if
all packages are successfully built first.  This is just a work around
until the order issues can be addressed, but it seems like a logical
thing to do in general.

Test Plan:

Deployed a snapshot.