[DWRF] Directly encoded ints optimization
Summary:
Analogous to my earlier diff "Directly encoded string optimization"

When the direct encoding for ints is carried over, we write the int to a long array first, and then to a stream to be compressed in the write stripe stage. The writing to the stream to be compressed can be done directly rather than going through the long array first. This also means that more data can be written into stripes since we store compressed data in memory, rather than the original uncompressed ints.

Test Plan: Ptests pass.