[DWRF] Avoid OOMs with many dynamic partitions/columns
Summary:
Sometimes queries will produce so many dynamic partitions with so many columns that the memory used just to initialize the writers will exceed the allocation given by the memory manager.  If this happens there's nothing the memory manager can do, and we tend to OOM.  Best case, we flush after every 10,000 rows.

To help avoid such cases, I added a check to the memory manager, so that if a Writer requests more than it's allocation just to initialize itself, dictionary encoding will be turned off on all writers (both new and old).  This greatly reduces the amount of memory required, to the point where we can generally avoid an OOM.  To compensate for the lack of dictionary compression, I also increased the Zlib level to 6 if we are using that compressor.

Unfortunately, when this happens ORC is slower than RC for writes.  To help compensate for this I added two optimizations.

If a writer flushes twice in a row on memory checks, and another writer does not flush for those two same memory checks, this implies that one writer is getting much more data than another.  To help with this I reallocate some memory from the unpopular writer to the popular writer, helping to reduce how often it needs to flush.

Also, when we flush the stripe, in the OutStream, if there is anything left in the buffer, we flush that to the OutputReceiver, and then immediately after flush the OutputReceivers data to disk.  We then have to reallocate buffers for all these streams when we start writing again.  To help with this, I allocated a buffer that can be shared across the streams for that list bit of data.  This greatly reduces the number of buffers that need to be allocated (e.g. present streams typically only need a single buffer) which helps with CPU.

It is still slower in the case I tried, so more work will be needed in the future.

Test Plan:
Ptests pass.  Added two new unit tests.

Tried writing to the 7 tables we saw fail in production, none of them experienced an OOM.