[DWRF] Don't read in all data at once
Summary:
Right now the way DWRF works is that it reads in all data for a stripe (minus excluded columns) from HDFS at once in a giant buffer and holds that in memory until the stripe is done being read.  The primary goals of this diff are to avoid this for 2 reasons:

1) We can null out buffers containing compressed index, dictionary data once they've been decompressed because we don't need the compressed data after this point.
2) We can only hold data for the current index stride (or a few strides) in memory (since this is the only data we'll be reading in the short term)

These should combine to make it DWRF much more memory efficient on read.

I will comment on this diff inline shortly (as there's a lot going on here).

Test Plan:
Ptests pass.

I did some performance testing too.

I tried preads instead of readfully, but readfully outperformed it, presumably because we read so much data in the same general area of the file.

I ran the perf tests, I haven't gotten conclusive results yet (I have to clean up some bad data first) but it looks like it's still at least 20% faster than RC (we were 25% faster originally).  For queries like count(*) we are significantly faster, I forget the exact number, but it's mostly due to the fact that we are reading a small fraction of the data from HDFS compared to before (I think just the stripe footers, dictionaries and indexes, this can be further optimized away in the future).  If we read e.g. every other column, we're about the same as before because in both cases we were reading a chunk of data for each column, rather than all the data as a giant chunk, which is similar to what we do now.  That 5% hit comes when we read all columns, this involves many more reads from HDFS than the single read from before, so this is expected, we make some of it back with fewer GCs, but not all of it.  I think in the future we can improve this, e.g. if we the nth stride for each column sequentially in the file we might get better performance (but there would be a cost since n strides for a single column would not be sequential).  Anyway, there's more work to do here.

In terms of memory, I was looking at a table with a lot of columns and would probably require a lot of memory (I don't know what table Presto was specifically having trouble with).  The amount of memory needed by the Reader dropped from ~43Mb to ~20Mb and could be as low as 18Mb adjusting configurations.  Tables with fewer columns should see a much more dramatic drop since they will have more index strides in a stripe (smaller rows = more rows per stripe).