[DWRF] Add config to start writing in low memory mode
Summary:
I got a task yesterday for a query that's failing with OOMs.

This query was writing a column with strings that could be > 3Mb.  DWRF cannot handle this well, mainly due to dictionaries and buffering of values.  If we turn on low memory mode immediately and turn off buffering the query can succeed.

In the future we can work on automatically detecting these, but this seems like a rare case and I didn't want to make queries less efficient by trying to detect these rare cases.

Test Plan: Running ptests.