[DWRF] bug in seek with chunks
Summary:
There is a bug (I think the comments in the test describe it well) where if you read from a stride in a chunk and then read from another stride at the same byte offset in another chunk the compressed offset will not be updated and you will get incorrect data.  This is because previousOffset will equal newCompressedOffset because both will be the byte offsets of the strides in the chunks.  We need to additionally check that we haven't read new data from disk (i.e. we aren't in a new chunk).

snapshottests

Test Plan:
Running ptests.

Verified it fixed the failing query in snapshot tests.

Added a very specific testcase.