[DWRF] Overflow error in large files
Summary: The InStream's offset is typed as an int which worked fine when we'd only pass in a byte array to it.  Now that we're passing the offset in the file, if the file is large enough, this can cause an error.  Changed the type to long.

Test Plan:
Added a testcase which simulates reading from a large file.

Ptests pass.