[DWRF] Make reading from HDFS lazily optional
Summary:
Reading lazily from HDFS causes a significant decrease in CPU performance, which is unavoidable as no matter how we arrange the data in the file some queries will always effectively do random reads in the file.

Adding an option to read lazily and defaulting it to off for now.  Applications strained for memory can turn it on by default if they choose, but for now, we are running into few enough issues (one application on one table) that it's probably not worth turning it on by default everywhere yet.

Test Plan:
Added a testcase.

ptests pass.