[DWRF] Dynamically turn on lazy HDFS reads if size is above threshold
Summary: Added a configurable threshold for the compressed size of the data to be read in by a DWRF RecordReader.  If the size of the data exceeds this threshold, it uses lazy reads to read that stripe, otherwise it performs the read eagerly.

Test Plan:
ptests pass

This fixed a few OOMs in snapshot tests (these must have been tests that were not converted to ORC the last time we rolled out, because they OOM with the currently deployed code as well).