[DWRF] Bug in StringTreeWriter across carry overs
Summary:
In the StringTreeWriter, if dictionary encoding is not used, and this is carried over, the dictionary will be nulled out to save memory.   If the number of strides between determining the encoding is exceeded, and then dictionary encoding is decided upon, the dictionary will be initialized without any parameters.

Before this just meant the data was always sorted, a bug, but it wouldn't lead to incorrect results.  Now however, the writer would expect the data to be sorted suitably for stride dictionaries (if the config was set) but the data would not be.  This led to incorrect results.

Test Plan:
Fixed the bug I saw in snapshot tests.

Added a few tests.  Only testStrideDictionariesWithoutStripeCarryover tests this bug.  The rest were tests I wrote in search of the bug, they seemed good, so I'm leaving them in.

ptests pass