[DWRF] Add close method to lazy tree readers
Summary:
One major source of OOMs in snapshot tests was the fact that when we start a new stripe or open a new file while reading DWRF data, there still maybe pointers to the previous row floating around (e.g. there's one in RecordReaderImpl itself).  This means that row can't be cleaned up from memory, and more importantly, the dictionaries, and other buffers it holds have to remain in memory.  So, as we read in a new stripe, since we unpack dictionaries eagerly, we have potentially two dictionaries per column in memory (plus some extra stuff).

When we start a new stripe, we must be in the next method, and the contract of this method means that it's safe to null out/clear all the structures the previous row pointed to.  Similarly, if we're opening a new file (within the CombineHiveRecordReader), close was called on the previous file, and the contract of close allows us to clear out all this memory.

Test Plan:
Ptests pass (note that many of the tests involve reading data from multiple stripes, so this code should be tested by the existing ones)

This fixes all but maybe one OOM in snapshot tests.