[DWRF] Add optional limit on the size of dictionaries
Summary:
Added a parameter that limits the size of the uncompressed dictionaries (by default to 100 Mb).  In OOMs we've been seeing, it's been pretty consistently because the size of the dictionaries are surprisingly large (e.g. 150-200 Mb).  These extreme sizes aren't strictly necessary for compression gains.  I converted all the inputs to snapshot tests using this and the total increase in size was about 0.002% (insignificant).

About 80 tables decreased in size, 481 remained unchanged, and 109 increased in size.  Of those 3 increased in size by more than 1% (1%, 3%, 4% respectively) on one decreased in size by almost 2%.

Given that this doesn't seem to have a significant effect on compression in most cases, and it could help with OOMs I think it's a reasonable trade off, and we can always turn it off on cold data or if we size a significant detriment.

Test Plan: ptests pass