[DWRF] Bug in low memory mode
Summary:
There's a bug in the string and integer tree writers when using the dictionary encoding, which only occurs when the writers enter low memory mode before the first index stride is written out.  If this happens, when the columns are converted from dictionary encoding to direct encoding, it will fail to create an index entry for the first index stride, which causes an IndexOutOfBoundsException later when the data is being read.

The fix just creates the row index entry if it's needed.

Test Plan:
Rewrote some partitions which ran into this issue and verified I could read them.

Added test cases.

Running ptests