[DWRF] Another bug in low memory mode
Summary: If a string column is written out directly and it's writing the stripe corresponding to the end of the hive.exec.orc.encoding.interval and at the same time it switches to low memory mode, it will not enter the if block to convert the dictionary encoded data to direct encoding.  This is because use dictionary encoding is false, but use carried over direct encoding is also false.  Due to this the data in the dictionary is lost.

Test Plan:
Added a unit test.

Ptests pass.