[DWRF] Yet another bug in low memory mode
Summary:
There is a bug in low memory where if low memory mode is entered in a stride (other than the first) and not enough rows are written to the strip to finish off this stride (e.g. if the number of rows in a stride is 10,000 rows, low memory mode is entered after writing 1000 rows to the stride and any number less than 9000 rows are written before the stripe is flushed) then the index entry will not be populated with the positions for that stride.

To fix this I added a check, if dictionary encoding is being used, and the number of saved indeces is > 0 then write out index entries for the number of saved indeces + 1.  This is because if we are in this case, we must be converting from a dictionary to direct encoding.  So write out an index entry for every stride we've finished so far, plus the one we're currently working on.

Test Plan:
Added a testcases to demonstrate these.  Also added a couple test cases for other edge cases I could think of.

ptests pass