[DWRF] Support getRecordWriter with null FileSystem
Summary: RC file and Sequence file both ignore the FileSystem parameter passed to the getRecordWriter method.  To make ORC more compatible with these old formats, check if the FileSystem parameter is null, and if so, use the RC file method of initializing it.

Test Plan:
Added a test case for it.

Ran ptests.