[DWRF] Materialize in the OrcLazyObject copy constructor
Summary:
We should call materialize in the OrcLazyObject copy constructor to prevent the copy and the original from both having materialize being called on them and both advancing the tree reader, also this prevents the copy from calling materialize out of order.  Added a comment to this effect in the code.

Also added a null check to all the copy constructors for classes which extend OrcLazyObject.  Previously, if materialize wasn't called previous would be null and this would cause an NPE.  Still, if the value of the first row is null, this could still be null, so we should check for it.

Test Plan:
Added unit tests.

Running ptests.