[DWRF] getRecordWriter backward compatible with RCFile
Summary:
getRecordWriter now checks if the work output path is set in the conf and prepends this to the output path if it is.

this, combined with the change earlier to get the filesystem from this same path, make it backward compatible with the way users are calling the same method on RCFile.

Test Plan: Added a test case to make sure it works, had a user verify it works for their use case.