[DWRF] Try flushing before closing
Summary:
We've seen a couple instances where the ORC file reader fails to parse the post script of a file (i.e. the last n bytes of the file).

So far I've been entirely unable to reproduce it.  This is my best guess for the cause (if we weren't flushing the last few bytes it would make the postscript invalid).

The HDFS team assures me this won't do anything, so worst case it has no effect.

RC file and Sequence file both call flush, so let's see if this has any effect.

Test Plan: ptests pass.